STYLE: removed warnings under win64 VS9

+++ Plugins/PointSprite/ParaViewPlugin/Qvis/GaussianControlPointList.cpp
@@ -597,7 +597,7 @@ GaussianControlPointList::RemoveControlPoints(int index)
 int
 GaussianControlPointList::GetNumControlPoints() const
 {
-    return controlPoints.size();
+    return static_cast<int>(controlPoints.size());
 }
 
 // ****************************************************************************
